requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.47.2
anyio==4.9.0
anytree==2.13.0
appnope==0.1.4
asgiref==3.9.0
astroid==3.0.3
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
bce-python-sdk==0.9.35
beautifulsoup4==4.12.3
boto3==1.34.162
botocore==1.34.162
cachetools==5.5.2
camel-converter==4.0.1
certifi==2025.6.15
cffi==1.17.1
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
comm==0.2.2
cryptography==43.0.3
curl_cffi==0.7.4
dashscope==1.19.3
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
Django==5.2.4
dnspython==2.7.0
et_xmlfile==2.0.0
executing==2.2.0
faiss-cpu==1.7.4
Farama-Notifications==0.0.4
fastjsonschema==2.21.1
fire==0.4.0
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
gitdb==4.0.12
gitignore_parser==0.1.9
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.40.3
google-generativeai==0.4.1
googleapis-common-protos==1.70.0
greenlet==3.2.3
grep-ast==0.3.3
grpcio==1.67.1
grpcio-status==1.62.3
grpcio-tools==1.62.3
gymnasium==0.29.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
htmlmin==0.1.12
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
imap-tools==1.5.0
ipykernel==6.27.1
ipython==8.17.2
ipywidgets==8.1.1
isodate==0.7.2
isort==5.13.2
jedi==0.19.2
jieba==0.42.1
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-path==0.3.4
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
lancedb==0.4.0
lazy-object-proxy==1.11.0
libcst==1.0.1
loguru==0.6.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
meilisearch==0.21.0
-e git+https://github.com/coljiang/MetaGPT.git@ffae922650aad915315a0626453fb38738ae35fe#egg=metagpt
more-itertools==10.7.0
motor==3.7.1
multidict==6.6.3
multiprocess==0.70.18
mypy_extensions==1.1.0
nbclient==0.9.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
openai==1.64.0
openapi-core==0.18.2
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
openpyxl==3.1.5
overrides==7.7.0
packaging==25.0
pandas==2.1.1
parse==1.20.2
parso==0.8.4
pathable==0.4.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
playwright==1.53.0
portalocker==2.10.1
prance==23.6.21.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.9.2
pydantic_core==2.23.4
pyee==13.0.0
PyGithub==2.6.1
Pygments==2.19.2
PyJWT==2.8.0
pylance==0.9.0
pylint==3.0.4
pymongo==4.13.2
PyNaCl==1.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.1
pyzmq==27.0.0
qdrant-client==1.7.0
qianfan==0.4.12.3
rank-bm25==0.2.2
ratelimiter==1.2.0.post0
redis==5.0.8
referencing==0.30.2
regex==2023.12.25
requests==2.32.4
retry==0.9.2
rfc3339-validator==0.1.4
rich==13.6.0
rpds-py==0.26.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3transfer==0.10.4
scikit-learn==1.3.2
scipy==1.16.0
semantic-kernel==0.4.3.dev0
semver==3.0.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.7
spark-ai-python==0.3.31
sqlparse==0.5.3
stack-data==0.6.3
ta==0.10.2
tenacity==8.2.3
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.7.0
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.66.2
traitlets==5.14.3
tree-sitter==0.23.2
tree-sitter-languages==1.10.2
tree-sitter-python==0.23.6
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.11.0
tzdata==2025.2
unidiff==0.7.5
urllib3==1.26.20
volcengine-python-sdk==1.0.130
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.15.0
yarl==1.20.1
zhipuai==2.1.5.20250701
